♻️ Refact: 네트워크 관련 폴더 및 파일 구조 정리
