+++ EatPic-iOS/Sources/Common/Protocol/Network/TokenProviding.swift
@@ -1,0 +1,14 @@
+//
+//  TokenProviding.swift
+//  EatPic-iOS
+//
+//  Created by jaewon Lee on 7/18/25.
+//
+
+import Foundation
+
+protocol TokenProviding {
+    var accessToken: String? { get set }
+    var refreshToken: String? { get set }
+    func refreshToken(completion: @escaping (String?, Error?) -> Void)
+}
