+++ EatPic-iOS/Sources/Core/Networks/AccessTokenRefresher.swift
@@ -1,0 +1,63 @@
+//
+//  AccessTokenRefresher.swift
+//  EatPic-iOS
+//
+//  Created by jaewon Lee on 7/18/25.
+//
+
+import Foundation
+import Alamofire
+
+class AccessTokenRefresher: @unchecked Sendable, RequestInterceptor {
+    private var tokenProviding: TokenProviding
+    private var isRefreshing: Bool = false
+    private var requestToRetry: [(RetryResult) -> Void] = []
+    
+    init(tokenProviding: TokenProviding) {
+        self.tokenProviding = tokenProviding
+    }
+    
+    func adapt(
+        _ urlRequest: URLRequest,
+        for session: Session,
+        completion: @escaping (Result<URLRequest, any Error>) -> Void
+    ) {
+        var urlRequest = urlRequest
+        if let accessToken = tokenProviding.accessToken {
+            urlRequest.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
+        }
+        completion(.success(urlRequest))
+    }
+    
+    func retry(
+        _ request: Request,
+        for session: Session,
+        dueTo error: any Error,
+        completion: @escaping (RetryResult) -> Void
+    ) {
+        guard request.retryCount < 1,
+              let response = request.task?.response as? HTTPURLResponse,
+              [401, 403, 404].contains(response.statusCode) else {
+            return completion(.doNotRetry)
+        }
+        
+        requestToRetry.append(completion)
+        if !isRefreshing {
+            isRefreshing = true
+            tokenProviding.refreshToken { [weak self] _, error in
+                guard let self = self else { return }
+                self.isRefreshing = false
+                
+                let result: RetryResult
+                if let error = error {
+                    result = .doNotRetryWithError(error)
+                } else {
+                    result = .retry
+                }
+                
+                self.requestToRetry.forEach { $0(result) }
+                self.requestToRetry.removeAll()
+            }
+        }
+    }
+}
