+++ EatPic-iOS/Sources/Core/Networks/DTO/UserInfo.swift
