--- EatPic-iOS/Sources/Core/Networks/TargetType/TargetTypePlaceholder.swift
@@ -1,2 +1,0 @@
-// make generate 생성시 빈 폴더를 Tuist가 인식할 수 있도록 하기 위한 임시 파일
-// 모든 폴더에 하나 이상의 파일이 생성된 후 삭제하면 됩니다.
